pyproject.toml
[tool.poetry]
name = "ahoinmobiliaria"
version = "0.1.0"
description = "Esta app es para el manejo de los inmuebles de AHOInmobiliaria"
authors = ["Jose Carlos Florez <aryeteinc@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = ">=5.2.3,<6.0.0"
pillow = ">=11.2.1,<12.0.0"
pymysql = ">=1.1.1,<2.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
django-tinymce = ">=4.1.0,<5.0.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"